Kit::Api::JsonApi - File structure refactoring
