+++ libraries/kit-api-json-api/app/services/kit/api/json_api/services/linkers.rb
@@ -1,0 +1,3 @@
+# Namespace for Linkers strategies.
+module Kit::Api::JsonApi::Services::Linkers
+end
