+++ libraries/kit-api-json-api/app/services/kit/api/json_api/services/resolvers.rb
@@ -1,3 +1,3 @@
-# Namespace for data Resolvers strategies
+# Namespace for data Resolvers strategies.
 module Kit::Api::JsonApi::Services::Resolvers
 end
