--- libraries/kit-api-json-api/app/services/kit/api/json_api/services/resolvers/data.rb
@@ -1,3 +1,0 @@
-# Namespace for Data Resolvers
-module Kit::Api::JsonApi::Services::Resolvers::Data
-end
