--- libraries/kit-api-json-api/app/services/kit/api/json_api/services/resolvers/serializers.rb
@@ -1,3 +1,0 @@
-# Namespace for Serializers
-module Kit::Api::JsonApi::Services::Resolvers::Serializers
-end
