+++ libraries/kit-api-json-api/app/services/kit/api/json_api/services/serialization.rb
@@ -1,0 +1,3 @@
+# Namespace for internal Serialization related logic.
+module Kit::Api::JsonApi::Services::Serialization
+end
