+++ libraries/kit-api-json-api/app/services/kit/api/json_api/services/serialization/query.rb
@@ -1,0 +1,39 @@
+# Serializer entry point.
+module Kit::Api::JsonApi::Services::Serialization::Query
+
+  include Kit::Contract::Mixin
+  # @hide true
+  Ct = Kit::Api::JsonApi::Contracts
+
+  before Ct::Hash[query_node: Ct::QueryNode]
+  after  Ct::Result[document: Ct::Document]
+  # Entry point to serialize a query AST
+  def self.serialize_query(query_node:)
+    Kit::Organizer.call({
+      list: [
+        self.method(:create_document),
+        Kit::Api::JsonApi::Services::Serialization::QueryNode.method(:serialize_query_node),
+      ],
+      ctx:  { query_node: query_node },
+    })
+  end
+
+  after Ct::Result[document: Ct::Document]
+  # Create a Document object that contains the json response and various caches
+  def self.create_document
+    document = {
+      cache:    {
+        resource_objects: {},
+        records:          {},
+      },
+      included: {},
+      response: {
+        data:     [],
+        included: [],
+      },
+    }
+
+    [:ok, document: document]
+  end
+
+end
