--- libraries/kit-api-json-api/app/services/kit/api/json_api/services/serializer.rb
@@ -1,4 +1,0 @@
-# Namespace for Serialization related logic.
-# TODO: remame this to `Serialization`
-module Kit::Api::JsonApi::Services::Serializer
-end
