+++ libraries/kit-api-json-api/app/services/kit/api/json_api/services/serializers.rb
@@ -1,0 +1,3 @@
+# Namespace for Serializers strategies
+module Kit::Api::JsonApi::Services::Serializers
+end
