--- libraries/kit-api-json-api/app/services/kit/api/json_api/services/sql/limit.rb
@@ -1,13 +1,0 @@
-# Generate LIMIT sql string statement.
-module Kit::Api::JsonApi::Services::Sql::Limit
-
-  include Kit::Contract::Mixin
-  # @hide true
-  Ct = Kit::Api::JsonApi::Contracts
-
-  before Ct::Hash[limit: Ct::Optional[Ct::Integer]]
-  after  Ct::Hash[sanitized_limit_sql: Ct::String]
-  def self.limit_to_sql_str(sorting:)
-  end
-
-end
