--- libraries/kit-api-json-api/app/services/kit/api/json_api/services/sql/sorting.rb
@@ -1,20 +1,0 @@
-# Transform sorting parameters from the AST to a SQL string
-module Kit::Api::JsonApi::Services::Sql::Sorting
-
-  include Kit::Contract::Mixin
-  # @hide true
-  Ct = Kit::Api::JsonApi::Contracts
-
-  before Ct::Hash[sorting: Ct::Optional[Ct::SortOrders], table_name: Ct::String]
-  after  Ct::Result[sanitized_sorting_sql: Ct::String]
-  def self.sorting_to_sql_str(sorting:, table_name:)
-    return true if !sorting || sorting.size == 0
-
-    str = sorting
-      .map { |column_name, sort_order| "#{ table_name }.#{ column_name } #{ sort_order.to_s.upcase }" }
-      .join(', ')
-
-    [:ok, sanitized_sorting_sql: str]
-  end
-
-end
